requirements.txt
fastapi[all]
SQLAlchemy
uvicorn
alembic
asyncpg
bcrypt
celery
fastapi-cache2
python-dotenv
redis
psycopg2
PyJWT
pillow
passlib
boto3
short_url
dependency-injector

